operations/svg/svg_huerotate.c operations/svg/svg_luminancetoalpha.c
2007-06-12  Mukund Sivaraman  <muks@mukund.org>

        * operations/svg/svg_huerotate.c
        * operations/svg/svg_luminancetoalpha.c
        * operations/svg/svg_matrix.c
        * operations/svg/svg_saturate.c: Use glong instead of gint.


svn path=/trunk/; revision=1564

+++ operations/svg/svg_huerotate.c
@@ -49,7 +49,7 @@ process (GeglOperation *op,
   const gchar  delimiter=',';
   const gchar *delimiters=" ";
   gchar      **values;
-  gint         i;
+  glong        i;
 
   m = ma;
