+++ operations/svg/svg_matrix.c
@@ -54,7 +54,7 @@ process (GeglOperation *op,
   const gchar  delimiter=',';
   const gchar *delimiters=" ";
   gchar      **values;
-  gint         i;
+  glong        i;
 
   m = ma;
